unify paths.py and pathlib.py
